Case Added
2 Cases added. In docker file commenting/uncommenting is required according to the case selected.
